Renames CH algorithms.
